don't use __builtin_frame_address on x86_64 architecture, it doesn't seem
2005-01-19  John Harper  <jsh@unfactored.org>

        * continuations.c (save_stack): don't use __builtin_frame_address
        on x86_64 architecture, it doesn't seem to give an address on the
        stack

        * lisp.c (Fbacktrace): fix long->int warning on 64 bit systems

        * sockets.c: use socklen_t instead of size_t where appropriate,
        fixes warnings on 64 bit systems

+++ src/lisp.c
@@ -2705,13 +2705,13 @@ ARGLIST had been evaluated or not before being put into the stack.
 		{
 		    char buf[256];
 #ifdef HAVE_SNPRINTF
-		    snprintf (buf, sizeof (buf), " at %s:%d",
+		    snprintf (buf, sizeof (buf), " at %s:%ld",
 			      rep_STR (rep_CAR (origin)),
-			      rep_INT (rep_CDR (origin)));
+			      (long) rep_INT (rep_CDR (origin)));
 #else
-		    sprintf (buf, " at %s:%d",
+		    sprintf (buf, " at %s:%ld",
 			     rep_STR (rep_CAR (origin)),
-			     rep_INT (rep_CDR (origin)));
+			     (long) rep_INT (rep_CDR (origin)));
 #endif
 		    rep_stream_puts (strm, buf, -1, rep_FALSE);
 		}
